migration for supabase

+++ app/models/inquiry.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String, func
+from sqlalchemy import BigInteger, DateTime, ForeignKey, String, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.core.database import Base
@@ -11,7 +11,10 @@ class Inquiry(Base):
 
     id: Mapped[int] = mapped_column(primary_key=True)
     reference_code: Mapped[str] = mapped_column(String(10), unique=True, nullable=False)
-    customer_telegram_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    # BigInteger: Telegram user IDs can exceed the 32-bit INTEGER range for
+    # newer accounts, which caused inserts to fail with "integer out of
+    # range" for some customers while working fine for older/smaller IDs.
+    customer_telegram_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     customer_username: Mapped[str | None] = mapped_column(String(60))
     product_id: Mapped[int] = mapped_column(ForeignKey("products.id"), nullable=False)
     preferred_color: Mapped[str | None] = mapped_column(String(40))
